Use AdaptedScriptLanguage where feasible
The JRuby and Jython scripting language implementations can wrap the
already-provided JSR-223-compliant ScriptEngineFactory instances.

The only thing not wrapped yet is the getScriptEngine() method itself
(we have a reimplemented ScriptEngine for each). We could partially
(mostly) reuse them too, but we will need to inject additional
functionality such as auto-importing, so the reuse cannot be total.