Change how scripts are discovered and managed
This is the start of a Grand Unification of the scripting framework, but
not yet complete. Previously, discovered scripts were Command wrappers
added to the CommandService using the ScriptPlugin adapter. However, a
script really is *not* a Command, because a Command is a Plugin, and
scripts are not Plugins (a Plugin is a Java class annotated with
@Plugin; scripts are not Java classes).

A script can, however, be expressed a Module. So we define a new type of
Module called ScriptModule to accomplish this, with a corresponding new
type of ModuleInfo called ScriptInfo for storing script-specific
metadata (in particular, the path to the script on disk, if any).