Add a new command for executing scripts
It is nice to have a Command that takes as input a file containing a
script to execute, and then executes said script (producing no formal
outputs). The ScriptPlugin was previously such a command, but it did
lean on ScriptService#eval, probably because that method did not yet
exist when it was first written.

The ScriptPlugin also doubled as a command wrapper for the discovered
scripts themselves, which was a rather unfortunate and ultimately
inflexible way of doing things (rather, we want each script to have
properly typed inputs and outputs to the extent possible).

So this change removes the old ScriptPlugin (nothing is using it anymore
anyway) and adds a simple RunScript command into ij-commands that
serves the former purpose, while abandoning the script wrapping aspect.