ScriptInfo: tweak ctor parameter order
We are going to add another constructor in a minute here, with a third
parameter, that "sort of" goes along with the "String path". To keep
those two next to each other, and to have the third parameter be third,
let's make the unrelated parameter (the Context) the first one.