ScriptInfo: add support for arbitrary Readers
There is now a constructor which allows passing a String (pseudo-)path
along with a Reader that provides the contents of the script. If a
Reader is given, that is used to parse inputs; if not, it is assumed
that the given path is an actual file on disk from which the contents
should be read.

It will typically be the case that ImageJ will discover scripts on disk
and wrap them in ScriptInfo instances using the File constructor. Such
scripts are the ones that will be managed by the ScriptService.

But running a script dynamically (e.g., from the Script Editor before
it is even saved) will be feasible by creating a new ScriptInfo object
which wraps the buffer itself in Reader form.