Move module processing plugins to subpackage
This renames the imagej.plugin package to imagej.module.process.
It also moves the imagej.module.Module*Processor classes into
imagej.module.process, as well.

It is probably slightly overengineered to keep the ModulePreprocessor
interface in addition to PreprocessorPlugin (and similar for
postprocessors), but the split exists for historical reasons. And the
separation of concerns there is actually kind of nice as well, since it
makes it rather simpler to implement a ModulePreprocessor if you don't
care about the discoverability granted by making it a plugin. That said,
we don't follow this pattern for any other ImageJ plugin types, so it is
perhaps a bit strange here. But we'll leave it this way, for now.