WIP: Wrap ImageJ 1.x plugins case by case
This approach has two major advantages:

1) The inputs & outputs of each individual legacy plugin become crystal
   clear, such that the plugin could potentially integrate with other
   software (KNIME, OMERO, etc.) the same as modern ImageJ2 commands.

2) Only images relevant to the plugin's inputs and outputs need to be
   translated during plugin execution. This narrowed focus represents a
   potentially huge performance improvement in some cases.

If this approach works well enough, we might even be able to drop the
previous LegacyCommand approach completely. While it would reduce the
extensibility of ImageJ2's backwards compatibility (a wrapper for each
legacy plugin would need to be manually created), it would simplify the
legacy layer and encourage developers to migrate plugins to ImageJ2.