ScriptService: update eval methods & rename to run
The new run methods now use ScriptInfo objects, and are patterned
after the existing run methods of ModuleService and CommandService.

Similarly, they return a Future<ScriptModule> rather than the script's
return value itself. (We can guarantee that the resultant Module will
always be a ScriptModule, even if the casting is a little ugly.)