Remove automatic SciJava context injection
This injection was a special case to provide scripts with access to the
SciJava context. But they can now easily have access to the SciJava
context simply by specifying an input parameter; e.g.:

    // @Context ijc

And this way, they can name the context variable whatever they want,
instead of it always being hardcoded to "IJ" which caused a name clash
with ImageJ1's ij.IJ class.