TextEditor: restore Python-specific behavior
The previous commit eliminated the Python-specific tab size reset logic.
It used to be that every time Python was selected as a language, or a
Python script was opened, tab size would be reset to 4; and for other
languages, it would be reset to 8.

This change restores the Python reset to tab width 4, but changes all
other languages to restore to the saved tab size preference instead.

Hopefully this is an acceptable short term compromise for the "Python
should be width 4" convention, while still enabling tab size to be
persisted as a preference in general.

[CTR: Original patch from Fiji at:
      https://github.com/fiji/fiji/pull/50]