JavaScriptScriptLanguage: implement decoding
The implementation is a little hairy because we don't necessarily have
direct compile-time access to the JavaScript Wrapper interface
implemented by the NativeJavaObject wrapper. But we can access it via
reflection. It is typically either org.mozilla.javascript.Wrapper, or in
the case of Oracle's shaded implementation,
sun.org.mozilla.javascript.internal.Wrapper. Either way, the package
will match that of the wrapped object itself, so we can go from there.