ConfigFileParameters: remove deprecated logic
The ConfigFileParameters class was performing operations that really
need an ImageJ context, but without having one. So now the class is
Contextual, making it much more straightforward for it to query the
ImageJ application version, use a proper LogService, etc.

The OptionsMemoryAndThreads plugin has been updated accordingly to
lazily pass the context to its ConfigFileParameters helper instance.