CommandFinderPanel: be defensive about file paths
It is possible that the associated file of a class can come back
null. It hasn't happened in practice before, but I encountered the
problem when I had stale SezPoz annotation metadata (thanks,
Eclipse!). In any case, it is better to be defensive, so that just
in case a weird situation comes up, the Command Finder UI will
simply show the file path as blank, rather than an NPE being thrown.