Do not use of Plugin#attrs for Tools
Instead, the Tool interface can simply have two API methods for querying
whether the tool is always active, and whether active in the app frame.

The main reason to use attrs is if there is metadata that must be known
*before* instantiating a plugin instance, or even before loading the
plugin class at all. But in the case of tools, they are singletons so we
always make one of each, which we can then interrogate as desired.