Move 'mvn exec:exec' support in app/ into a profile
Defining the configuration outside any profile makes it impossible
for profiles to disable the exec:exec phase otherwise. And we will
have to disable this in the 'deps' profile when the annotation
combiner is to be run as a part of the assembly.

So let's attach the exec:exec goal to the test phase if, and only if,
the exec profile is active.

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>