Use the shiny new META-INF combiner for uber jar files
We need to combine more than just the annotation indexes: script engines
implementing the javax.scripting interface need to be declared in
META-INF/services/ for example.

Hence we introduced the META-INF combiner in SciJava-commons whose job
it is to know all the things that need to be combined and just do that.

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>