AdaptedScriptLanguage: fail early when passed a null ScriptEngine
Instead of failing later, when the script engine is used, we should fail
early to mark this as an invalid adapted script language.

This can happen when, say, an old JRuby is installed (as is currently the
case in Fiji) that does not yet offer a javax.script service via
META-INF/services/.

Without this fix, the script engine would be installed and later would
prevent proper startup of a required service, rendering the complete
context invalid.

Identified by Mark Hiner as the problem underlying the bug "Fiji will not
quit" reported by David Knecht.

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>