check-headers.pl: fix copyright header checking
The LICENSE.txt file is no longer 100% identical to the header at
the top of each source file. As such, we cannot rely on an exact
comparison between the two. So let's relax the check to merely verify
that some sort of copyright header comment is present at the top.

This is OK, since running "mvn license:update-file-header" will
steamroll over any errors in such header comments anyway.