Update project licenses
This was done using:
  mvn license:update-project-license

The only real change is the loss of the legal disclaimer at the bottom
of the license. This disclaimer is not actually part of the BSD-2
license proper, but rather an addendum. For relevant discussion, see:
  https://jira.codehaus.org/browse/MLICENSE-80

See also dcb4444d063efb4c1e707edf514f2ff7779dbb42.