Remove deprecated run methods
These were put in place before the 2.0.0-beta-7.6 release, so that
downstream code had a clear migration path (despite the fact that
ImageJ2 is still in beta, we try to keep old API in place for one
release as deprecated before removing).

However, in this case, due to method signature ambiguities, it is
important that the deprecated methods be removed ASAP since they
unfortunately interfere with use of the new method signatures.