Use ImageDisplay-specific context menu
For the context menu items intended to be used with ImageDisplays,
let's use a more specific menu root string.

This change is in preparation for
org.scijava.plugin.Plugin#CONTEXT_MENU_ROOT being deprecated in the
next release of scijava-common.

Now, it may be the case that this whole idea of "context menu roots"
doesn't work well once we flesh out various context menus further.
I have my doubts. But we'll give it a try first (iterate faster!)
and see how it goes.