ScriptFinder: improve script discovery
We remove the "special Scripts subdirectory" logic in favor of
consistent behavior across all script candidate directories.

Scripts in the toplevel directory are now ignored, rather than being
placed into the Plugins menu directly. The main reason for this is
backwards compatibility: we want to remain compatible with Fiji's
existing "scripts" directory which currently contains .m files which
should *not* be discovered and placed into the main menu structure, but
rather exist for use with Miji.