DefaultScriptService: fix lazy init race condition
For performance, we use an unsynchronized null check against each data
structure in order to determine whether to enter the synchronized
initialization method for that data structure. Once inside, the null
check is repeated in case the data structure was initialized in the
meantime by a different thread.

The previous logic was flawed though, because after confirming that the
data structure was still null, it immediately _initialized that data
structure's object reference_ before completely initializing it. The
correct behavior is to first fully initialize the data structure in a
local variable, and only at the very end assign the correct pointer to
the data structure's object reference. Otherwise, a race condition
results when two threads call the data structure's getter method
near-simultaneously: if a first thread begins initializing, then a
second thread queries the getter and sees that the object reference is
non-null, that reference is immediately returned even though the first
thread is still not done initializing the structure itself.