ScriptInfo: parse script parameters lazily
This change adds a simple callback method in the abstract layer which
concrete subclasses can override to define how parameters are parsed
upon first access.

This lazy parsing reduces the startup time of ImageJ by allowing scripts
to appear in the menu structure without needing to be parsed first.

For now, we keep ScriptInfo's parseParameters() method at public
visibility, since it is nice to be able to reparse parameters in case
the script changed.