RELEASES - Consolidate PREREQUISITES instructions
As all prerequisite components are now deployed basically the same as each other (through release-version.sh), there is not really any reason to have separate sections for each. So these instructions have been consolidated.

As each component's instructions had boilerplate mentioning the bump-pom-scijava usage, this was split out to its own section detailing what to do with pom-scijava when there are breaking changes or not.