Discover script editor templates in script-templates/ directories
Whenever a class path element contains the directory script-templates/,
the script editor will enumerate all files within and interpret the
relative paths as menu paths in the Template menu. When the user selects
the respective menu entries, the respective file will be loaded into the
script editor, the base name set, and the language chosen according to the
recorded file extension.

Note: Typically, .jar files built by Maven or Ant will have extra entries
for directories. These are not strictly necessary for using the classes
(as the .class entries are accessed directly) but they are necessary for
the script template discovery to work. Old Fiji Build and older versions
of MiniMaven are known to produce incorrect .jar files; don't use them.

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>