Close the script's output streams only after the script is done
It was not quite clear from the contract implicitly defined by the name
'evalScript()' that it is a fire-and-forget method; 'eval' usually
implies that the respective entity is to be evaluated completely. This
is fixed now: evalScript() will return only when the script execution
has finished.

Implicitly, this also fixes the problem where evaluating a single-liner
Beanshell script printing a simple string would close the output streams
before the string was printed.

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>