No longer inject the Context & LegacyService in ij.IJ
We cannot: the class loader in which ImageJ 1.x lives should not need
to know any ImageJ2 class explicitly.

Instead, calling

	ij.IJ.runPlugIn("org.scijava.Context", null);

will be intercepted by the legacy hooks and ask the LegacyService itself
for the context.

Hence we no longer need to inject the Context nor the LegacyService into
the ij.IJ class as static fields.

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>