LegacyInjector: set class loader on new context
When we create the IJ2 context, we want to propagate the context class
loader from the calling thread. This will help IJ2 use IJ1's plugin
class loader while it is booting up. This is important because upon
completion of the SciJava context services being loaded, a
ServicesLoadedEvent is fired using EventService#publishLater which calls
ThreadService#run which creates a new thread in the thread pool (since
it was the first time run has been called). And it is paramount that
each and every thread of the ThreadService's thread pool gets its
context class loader set to the IJ1 plugin class loader, or else any
code later executed via ThreadService#run may not have access to all the
classes loaded via that class loader. In particular, commands executed
through the IJ2 UI such as "System Information" may not have access to
all classes and resources made available via the plugin class loader.