Move installHooks(LegacyHooks) from CodeHacker to LegacyInjector
The installHooks() method is really meant to run long, long, long time
after the code patching happened. Therefore, the CodeHacker class is
actually not the best place to put that method (but we did, in order to
be certain to insert the hooks into the correct class loader).

While at it, get rid of the setLegacyService() method: we do not set
the legacy service at all any longer, just the legacy hooks: it is their
task to connect everything to the correct legacy service.

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>