CodeHacker: add a way to commit non-IJ1 classes into the given class loader
We add three classes in the legacy patching: LegacyHooks,
EssentialLegacyHooks and HeadlessGenericDialog. These must be known to
the class loader, of course.

With LegacyHooks, we want to make it accessible both to the calling
classes and ImageJ, so it makes sense to share that interface (via the
slightly method to return it from findClasses(), violating the
convention that a class loader should only return classes whose
associated class loader is in its hierarchy).

With EssentialLegacyHooks and HeadlessGenericDialog, this is not
possible: they access the ImageJ 1.x classes themselves, and therefore
they would access the *wrong* ImageJ 1.x classes if we were to reuse
those classes from a different class loader. Therefore, we need to
commit those classes to the same class loader the ImageJ 1.x classes
are living in.

This makes only a difference if the class loader that gets passed to
CodeHacker is a different one than the one which loaded CodeHacker.
That is exactly the way we intend to handle things in the near future:
to be at least able to encapsulate ImageJ 1.x "instances" by loading
them into special-purpose class loaders.

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>