Avoid using ImageJ 1.x classes in LegacyHooks directly
Since we want to share the LegacyHooks interface between the
LegacyClassLoader and the ClassLoader that performed the ImageJ 1.x
patching, the LegacyHooks interface cannot pass an ImagePlus
instance around: it would be incompatible (by reference, the
LegacyHooks' reference to ImagePlus would be loaded from the class
loader that loaded the LegacyHooks interface itself).

Side-step the whole issue by passing Object instances around, not
ImagePlus instances.

This also opens the door for the future when we might want to have a way
to translate between ImagePlus instances loaded into different class
loaders.

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>