Add two more methods to the legacy hooks
The whole idea of the legacy patching is to provide more extension
points than ImageJ 1.x offers on its own.

One very important extension point required by literally every
application that might need to make use of extension points is a
callback when ImageJ 1.x was initialized.

There are possible chicken-and-egg problems here, with the legacy hooks
living in ImageJ 1.x' jars/ directory: the legacy hooks can only be
loaded using ImageJ 1.x' PluginClassLoader, therefore they cannot be
installed before ImageJ 1.x was initialized.

To remedy that situation, we change the EssentialLegacyHooks (i.e. the
hooks installed initially) to interpret the system property
ij1.patcher.initializer as name of a class implementing the Runnable
interface, to be called upon initialization.

Of course, this is too late for some of the extension points (such as
setting the window title of the ImageJ 1.x main window), but that cannot
be helped: we simply have no way to access classes in the
PluginClassLoader before that class loader's class path is determined --
which happens *during* the initialization of the main window, i.e.
*after* the window title/icon are set.

It is therefore the duty of the initializer to fix that, first thing
(unless we are running in batch mode, in which case IJ.getInstance()
will be null).

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>