Ensure initial legacy hooks are loaded using the correct class loader
The class loader used to load CodeHacker might be different from the one
into which we want to load the patched ImageJ 1.x. To ensure that the
initial legacy hooks can find the correct ImageJ, we must therefore load
the initial legacy hooks using the same class loader as ImageJ 1.x.

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>