Install the initial legacy hooks in a class initializer
This is more robust, e.g. when writing out a headless.jar; In such a
case, there is no runtime patching, and therefore no other way to
initialize the hooks than to put that code into the class initializer
(suggesting that this is the proper way to do it anyway, separating
the concern of installing the hooks from patching the code).

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>