Allow to set the macro options in a LegacyEnvironment
When calling run() or runMacro(), one can specify the macro options as a
second parameter. However, when calling runPlugIn(), the second
parameter specifies the 'arg' parameter passed to the plugin's run() or
setup() method.

In this commit, we introduce the 'setMacroOptions()' method that allows
to set the macro options even for runPlugIn() calls.

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>