Fix getDefaultUI() for lazy initialization
It used to be the case that uiList was initialized at service
initialization time. This is no longer true: it is initialized lazily,
i.e. when uiList() is called. Therefore it is no longer good enough to
test *before* looking at the uiList whether there is a defaultUI, but we
need to test also afterwards.

This is most likely the culprit behind the issue reported in
https://github.com/imagej/imagej/issues/21 and behind the second issue
reported in https://github.com/imagej/imagej/issues/26.

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>