Intermediate release of ij-legacy
A lot of things happened in ij-legacy recently. To provide a smooth
upgrade path for Fiji, let's deploy an intermediate release ij-legacy so
that fiji-compat can already support the new startup strategy while
falling back to the old one. In particular, Fiji will need to implement
the plugin types defined in imagej.legacy.plugin.*.

This change makes an intermediate release of ij-legacy, beta 7.7.1, that
relies on beta 7.7 of the rest of ImageJ2.

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>