Remove dependence on OptionsService and ImageDisplay
ActiveImagePreprocessor now handles Position, Dataset, and
ChannelCollection inputs, filling with default values if a single input
is needed. Note the TODO as preprocessing a ChannelColleciton input
should likely be done elsewhere but unsure where.

FillDataValues: Now longer relies on the OptionsService to get a color
for filling with. It also no longer relies on ImageDisplay. Instead it
takes, a ChannelCollection, Dataset, and Position.

Services: The Display and Overlay Services (Interfaces and default
implementations) were changed to accommodate the need to access the
active dataset and active position. Largely a suggestion but I expect
to use the methods widely in many other plugins.