Tweak SWT and Pivot GAVs when ij.ui is set
The artifact IDs changed to use the "imagej-" prefix. The group ID
changed to {$imagej.groupId} to match other ImageJ projects, rather than
relying on ${project.version} to necessarily match (it will here, but
might not if someone copy/pasted the dependency block to elsewhere).
The version is now omitted with the idea that it will be managed in the
future. In the meantime, setting ij.ui to pivot or swt won't work, but
those interfaces are not yet functional anyway.