Do not let managed imagej.version take precedence
Rather, each componenet should depend on the current working copy's
version of each module in the multi-module build (i.e., on
${project.version}, not the ${imagej.version} declared in pom-scijava).

However, managed versions always take precedence. So it is not enough to
declare '<version>${project.version}</version>' as that will be
overridden by the dependencyManagement of pom-scijava. Instead, we
override the ${imagej.version} property to equal ${project.version} so
that the current multi-module build's version *is* the managed version.
This has the nice side effect of letting us remove the '<version>' tags
altogether from all project dependencies.