Add service which checks for updates on startup
Previously, this was hard-coded into the UIService, which was the wrong
place for it. Now that the core UI subsystem has moved to SciJava
Common, let's take the opportunity to perform the update checking in a
better way: using a dedicated UpdateService. Over time, we can migrate
core updater functionality to this service, too.