Fix generation of uberjar
Finally it uses the Maven assembly plugin to do *all* the work,
including bundling all the sources.

There *might* be a problem with using
net.imagej:imagej:${project.version}:sources as a dependency (it might
grab the previous version of the sources from ~/.m2/repository; I'm not
sure), but I don't care right now. This works, because Jenkins always
builds and deploys first, *then* tries to generate the uberjar in a
separate execution.