Add toplevel ImageJ app
This overrides the one declared in imagej-common, so that the version
reported in the ImageJ application will reflect the version of
net.imagej:imagej rather than that of net.imagej:imagej-common.

It also reports the version of ImageJ 1.x, when available, in
parentheses, the way that imagej-legacy's LegacyImageJApp used to do.