POM: bump to pom-imagej 2.14
This bumps the imagej-launcher to version 3.0.2, which fixes a bug
where the launcher fell back to legacy mode (ij.ImageJ) instead of
using net.imagej.Main as expected.

It also bumps imagej-ui-swing to 0.3.1, which adds imagej-ui-awt as
a runtime dependency. This is needed so that text rendering works.

Lastly, it bumps some components which improve scripting support and
fix a few other bugs:
* scijava-common -> 2.19.1
* minimaven -> 2.1.0
* scripting-java -> 0.2.0
* scijava-ui-swing -> 0.2.0
* ij1-patcher -> 0.4.3
* imagej-legacy -> 0.3.0