POM: only create application ZIP if -Papp is set
We do not want to deploy the application ZIP to our Maven repository
anymore. It is large, and we are planning to cut a new release of
net.imagej:imagej whenever _any_ of its components are updated.

So let's hide creation of the application ZIP behind a profile, and let
Jenkins build it without deploying it anymore.