ToplevelImageJApp: use version from LegacyService
The LegacyService now has a getCombinedVersion() method, which is
exactly what we need for the toplevel ImageJ application.

However, there is currently an issue where the LegacyService is not
injected properly into the ToplevelImageJApp. This is because during
LegacyService's initialize method, it calls IJ.showStatus which triggers
a chain of events culminating in initialization of the default app.
Once we resolve that problem, though, then this change will matter.