Bump parent to 2.53
Unfortunately, this is the first time that the maven-enforcer-plugin
really had a chance to attack our dependencies, therefore we have to add
a really big maven-enforcer-plugin section to make it compile again.

Of course these configurations need to go into the relevant POMs, which
we will do after releasing rc-16.

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>