POM: put all UI plugins on the classpath
In general, it is bad form to manipulate the dependencies in profiles.
We were doing it to avoid triggering UI-activation-related bugs, but
actually that is why we _should_ have all the UIs on the classpath: to
quickly identify and squash any bugs that crop up when multiple UIs are
present on the classpath.