Add completeness test for ImageJ services
This test mostly exists to catch services which should implement
ImageJService, but don't. The current (known) offender is LegacyService,
fixed on the master branch, but not yet released. When that release
happens and this component is updated, the test will "break" and we can
uncomment the LegacyService.class line as originally intended.