POM: pin some versions to maintain Java 6 support
This commit is unstable, intended to be continually rebased on top of
the latest master branch. In this way, the branch provides the newest
possible combination of components that continue to work with Java 6.

It will never be released, and is provided mainly to document the
component versions that currently ship on the core ImageJ update site
(http://update.imagej.net/) until the transition to Java 8 is done.