Make imagej-legacy optional
This means it will no longer be inherited as a transitive dependency
by downstream projects. It must be included explicitly.

This also removes the legacy() accessor from the ImageJ gateway,
to avoid related class loading issues when imagej-legacy is absent.

For this to fully work, the LegacyService in imagej-legacy needed to
be updated to implement the Versioned interface. See:

https://github.com/imagej/imagej-legacy/commit/1a6aa0cde194cea697ef5af4eb210dd93ba4d081

This commit is dedicated to Kevin Mader!