Test that all core services are marked properly
Every core service should be _either_:

1) A SciJavaService
2) An ImageJService
3) A SCIFIOService

And not more than one of these.

It is a common error that new services are introduced which are not
properly marked with the appropriate interface. This causes them to
then be left out of the corresponding Gateway object.

For example, as of this writing, net.imagej.ops.NamespaceService does
not implement ImageJService as it should, and hence the service is not
available when creating a context via "new net.imagej.ImageJ()", since
the gateway (by default) includes only services of the above types.

This test will ensure that such errors are caught much sooner.