Ignore, rather than warn, about empty code blocks
Empty code blocks come up all the time, and it is annoying to have to
always document them. In particular, we are going to introduce a Nil
object in SciJava Common for use with ImageJ Ops, which allows for
generically typed nulls, but in order for it work, you must declare
an anonymous subclass so that the generics information is saved; e.g.:

    nilImg = new Nil<Img<T>>() {};

Needing to suppress the warning for each and every time we write this
invocation would uglify the code quite a lot.