POM: update parent to pom-scijava 24.2.0
This updates the following dependencies:

* imagej-common: 0.28.0 -> 0.28.1
* scijava-search: 0.5.0 -> 0.6.0

The imagej-common update simply brings in newer transitive dependencies.
For those importing the pom-scijava bill of materials, this does not
matter. But for those depending on ImageJ directly without using the
BOM, it affects which version of the scijava-table component (and
versions of other components) gets selected by Maven.

I consider this to be a bug in Maven; see:

* https://stackoverflow.com/q/45041888/1207769
* https://issues.apache.org/jira/browse/MNG-5761
* https://gist.github.com/ctrueden/d058330c8a3687317806ce8cc18332c3
* https://github.com/scijava/jgo/issues/9