ImageJ logo update (#252)
Created a new logo based on Jan Eglinger's submission in the logo
contest. Color inspiration was drawn from the second place winner, Dave
Williamson. Included is several sizes of the logo "flat" and with a
"shadow". Source files (created in inkscape) are also included. Please
view the readme file for specifics.