Update developers and contributors
Previously, this section described only the people related to the
net.imagej:imagej component specifically. But this component is used for
the statbox at https://imagej.net/software/imagej2 -- and it's the
blanket component bringing in the various ImageJ2 dependencies in one
convenient place -- so it makes sense to give credit to everyone who has
worked on ImageJ2 in general, not only this one tiny toplevel component.
So I pulled this list together from the POMs of the various imagej-*
components.