POM: purge imagej-launcher dependency
And update the POM's parent to pom-scijava 42.0.0.

The imagej-launcher component is not necessary to run ImageJ2.
Furthermore, the Fiji distribution of ImageJ(2) has switched to the
SciJava App Launcher (https://github.com/scijava/app-launcher),
which is used at runtime to do startup-related things such as
show a splash screen and check the version of Java being used.
But none of that is strictly needed to run the application.
E.g. Maven-aware deployments can simply use the net.imagej:imagej
dependency and run the net.imagej.Main class's main method.